renamed file_ensurer and score_rater
